Moved widget code to own subpackage. Widget seems not very functional (connection detection problems).
